Project/03_coffee_shop_full_stack/starter_code/backend/requirements.txt
astroid==2.12.2
cffi==1.17.1
click==7.1.2
colorama==0.4.6
cryptography==43.0.1
ecdsa==0.19.0
Flask==2.0.0
Flask-Cors==4.0.1
Flask-SQLAlchemy==2.5.0
future==0.17.1
greenlet==3.0.3
isort==4.3.18
itsdangerous==2.2.0
Jinja2==3.1.4
lazy-object-proxy==1.4.0
MarkupSafe==2.1.5
mccabe==0.6.1
pyasn1==0.6.0
pycparser==2.22
pycryptodome==3.20.0
pylint==2.3.1
python-jose==3.3.0
rsa==4.9
six==1.16.0
SQLAlchemy==1.4.41
Werkzeug==2.0.0
wrapt==1.11.1